feat: finalize test infrastructure improvements (#47)
- Update pre-commit message to reference nanostore/testutil/model_test.go
- Update technical architecture docs with fixture-based testing strategy
- Rename all _migrated_test.go files to standard _test.go names

This completes the test infrastructure refactoring, providing:
- Fixture-based testing with universe.json
- Clear test documentation via model_test.go
- Test categorization with appropriate header comments
- Developer guidance through pre-commit hooks
